requirements.txt
mcp
langchain-mcp-adapters
langchain-openai
python-dotenv
flask
markdown
